tools/ascii_resizer/pyproject.toml
[project]
name = "ascii-resizer"
version = "0.1.0"
description = "Resize and standardize ASCII art to consistent dimensions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["click>=8.0", "numpy>=1.24", "pillow>=10.0"]

[project.scripts]
ascii-resizer = "ascii_resizer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ascii_resizer"]

[tool.hatch.build.targets.sdist]
include = ["src/ascii_resizer"]

[tool.hatch.build]
sources = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=9.0.2"]
